requirements.txt
Django==1.11.16
django-cors-headers==2.2.0
djangorestframework==3.9.0
pkg-resources==0.0.0
pytz==2018.7
gunicorn==19.6.0
whitenoise==3.2.1
psycopg2==2.7.4
psycopg2-binary